Decoupled environment setup from playwright
